+++ HousingSearchApi/V1/Interfaces/IQueryFactory.cs
@@ -1,0 +1,9 @@
+using HousingSearchApi.V1.Boundary.Requests;
+
+namespace HousingSearchApi.V1.Interfaces
+{
+    public interface IQueryFactory
+    {
+        IQueryGenerator<T> CreateQuery<T>(HousingSearchRequest request) where T : class;
+    }
+}
